refactor: date-ranges-view
